Move away un-routable scenes to appropriate places
Scenes in this architecture should consist of only routable
components for consistency. Modal, Notification and Header have no
place in scenes according to this criteria.

The Header scene is removed and the wirings it did are moved to
parent scenes. They will eventually be abstracted in a layout
component.
